+++ src/main/java/domain/ladder/LineFloor.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 public record LineFloor(int value) {
     private static final int MINIMUM_LINE_FLOOR = 0;
