+++ src/test/java/domain/ladder/LineNumberTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.ladder;
 
+import domain.ladder.LineNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
